#104 refactor: LabelTag props 수정
